requirements.txt
tensorflow
tensorflow_hub
requests
numpy
Pillow
grpcio
tensorflow_serving_api